fixed: tooltips showing latex error message not correctly displayed on zooming in/out.

+++ LaTeXDraw/src/net/sf/latexdraw/glib/views/Java2D/impl/LTextView.java
@@ -468,7 +468,7 @@ class LTextView extends LShapeView<IText> implements IViewText {
 
 
 	@Override
-	public boolean isToolTipVisible(int x, int y) {
+	public boolean isToolTipVisible(double x, double y) {
 		return border.contains(x, y);
 	}
